Implement dvb_sub_open_pid, dvb_sub_close_pid and dvb_sub_read_data.
Also change open_pid to take guint16 for pid, not guint32, as MPEG-TS PIDs are 16bits (or less?)

+++ src/dvb-sub.h
@@ -123,7 +123,7 @@ GType    dvb_sub_get_type      (void) G_GNUC_CONST;
 DvbSub  *dvb_sub_new           (void);
 gint     dvb_sub_feed          (DvbSub *dvb_sub, guint8 *data, gint len);
 gint     dvb_sub_feed_with_pts (DvbSub *dvb_sub, guint64 pts, guint8 *data, gint len);
-int      dvb_sub_open_pid      (DvbSub *dvb_sub, guint32 pid);
+int      dvb_sub_open_pid      (DvbSub *dvb_sub, guint16 pid);
 void     dvb_sub_close_pid     (DvbSub *dvb_sub);
 void     dvb_sub_read_data     (DvbSub *dvb_sub);
 void     dvb_sub_set_callbacks (DvbSub *dvb_sub, DvbSubCallbacks *callbacks, gpointer user_data);
